CRUDFiltring/server/requirements.txt
websockets==11.0.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0 